requirements.txt
aiopubsub-py3
sanic
redis